BOAC-5111, v-dialog for AdvSearch; improve structure of header components
