+++ src/api/api-utils.ts
@@ -1,20 +1,20 @@
-import _ from 'lodash'
+import {each, get, isNil} from 'lodash'
 import axios from 'axios'
 import {useContextStore} from '@/stores/context'
 
 export default {
-  apiBaseUrl: () => _.get(useContextStore().config, 'apiBaseUrl'),
+  apiBaseUrl: () => get(useContextStore().config, 'apiBaseUrl'),
   postMultipartFormData: (
     path: string,
     data: object
   ) => {
     const formData = new FormData()
-    _.each(data, (value, key) => {
-      if (!_.isNil(value)) {
+    each(data, (value, key) => {
+      if (!isNil(value)) {
         formData.append(key, value)
       }
     })
-    const apiBaseUrl = _.get(useContextStore().config, 'apiBaseUrl')
+    const apiBaseUrl = get(useContextStore().config, 'apiBaseUrl')
     return axios
         .post(
             `${apiBaseUrl}${path}`,
