+++ src/api/reports.ts
@@ -1,4 +1,4 @@
-import _ from 'lodash'
+import {replace} from 'lodash'
 import axios from 'axios'
 import utils from '@/api/api-utils'
 
@@ -9,7 +9,7 @@ export function getAvailableDepartmentReports() {
 
 export function downloadAlertsCSV(fromDate: string, toDate: string) {
   const fileDownload = require('js-file-download')
-  const filename = `boa-alerts-${_.replace(fromDate, '/', '-')}-to-${_.replace(toDate, '/', '-')}`
+  const filename = `boa-alerts-${replace(fromDate, '/', '-')}-to-${replace(toDate, '/', '-')}`
   return axios
     .post(`${utils.apiBaseUrl()}/api/reports/download_alerts_csv`, {
       fromDate,
